+++ database/migrations/2020_11_06_042340_create_harvest_maintain_table.php
@@ -19,13 +19,13 @@ class CreateHarvestMaintainTable extends Migration
             $table->foreign('rkh_maintain_id')->references('id')->on('rkh_maintains')->onDelete('cascade');;
             $table->foreignId('employee_id')->constrained();
             $table->integer('amount_used');
-            $table->decimal('coverage', 8, 2);
+            $table->float('coverage', 8, 2);
             $table->string('image')->nullable();
             $table->date('date');
             $table->time('maintain_time_start')->nullable();
             $table->time('maintain_time_end')->nullable();
-            $table->decimal('lat', 10, 8)->nullable();
-            $table->decimal('lng', 11, 8)->nullable();
+            $table->float('lat', 10, 8)->nullable();
+            $table->float('lng', 11, 8)->nullable();
             $table->timestamps();
         });
     }
