+++ database/migrations/2020_11_06_042436_create_manual_maintain_table.php
@@ -19,15 +19,15 @@ class CreateManualMaintainTable extends Migration
             $table->foreign('rkh_maintain_id')->references('id')->on('rkh_maintains')->onDelete('cascade');;
             $table->foreignId('employee_id')->constrained();
             $table->integer('circle');
-            $table->decimal('circle_coverage', 8, 2);
+            $table->float('circle_coverage', 8, 2);
             $table->integer('pruning');
-            $table->decimal('pruning_coverage', 8, 2);
+            $table->float('pruning_coverage', 8, 2);
             $table->integer('gawangan');
             $table->date('date');
             $table->time('maintain_time_start')->nullable();
             $table->time('maintain_time_end')->nullable();
-            $table->decimal('lat', 10, 8)->nullable();
-            $table->decimal('lng', 11, 8)->nullable();
+            $table->float('lat', 10, 8)->nullable();
+            $table->float('lng', 11, 8)->nullable();
             $table->timestamps();
         });
     }
